Move logic to app package
