+++ app/import_functions.go
@@ -339,7 +339,7 @@ func (a *App) importUser(data *UserImportData, dryRun bool) *model.AppError {
 	} else {
 		var err error
 		// If no AuthData or Password is specified, we must generate a password.
-		password, err = model.GeneratePassword(*a.Config().PasswordSettings.MinimumLength)
+		password, err = generatePassword(*a.Config().PasswordSettings.MinimumLength)
 		if err != nil {
 			return model.NewAppError("importUser", "app.import.generate_password.app_error", nil, err.Error(), http.StatusInternalServerError)
 		}
