+++ model/user.go
@@ -4,13 +4,11 @@
 package model
 
 import (
-	"crypto/rand"
 	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math/big"
 	"net/http"
 	"regexp"
 	"sort"
@@ -908,53 +906,4 @@ func UsersWithGroupsAndCountFromJson(data io.Reader) *UsersWithGroupsAndCount {
 	bodyBytes, _ := ioutil.ReadAll(data)
 	json.Unmarshal(bodyBytes, uwg)
 	return uwg
-}
-
-var passwordSpecialChars = "!$%^&*(),."
-var passwordNumbers = "0123456789"
-var passwordUpperCaseLetters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-var passwordLowerCaseLetters = "abcdefghijklmnopqrstuvwxyz"
-var passwordAllChars = passwordSpecialChars + passwordNumbers + passwordUpperCaseLetters + passwordLowerCaseLetters
-
-func randInt(max int) (int, error) {
-	val, err := rand.Int(rand.Reader, big.NewInt(int64(max)))
-	if err != nil {
-		return 0, err
-	}
-	return int(val.Int64()), nil
-}
-
-func GeneratePassword(minimumLength int) (string, error) {
-	upperIdx, err := randInt(len(passwordUpperCaseLetters))
-	if err != nil {
-		return "", err
-	}
-	numberIdx, err := randInt(len(passwordNumbers))
-	if err != nil {
-		return "", err
-	}
-	lowerIdx, err := randInt(len(passwordLowerCaseLetters))
-	if err != nil {
-		return "", err
-	}
-	specialIdx, err := randInt(len(passwordSpecialChars))
-	if err != nil {
-		return "", err
-	}
-
-	// Make sure we are guaranteed at least one of each type to meet any possible password complexity requirements.
-	password := string([]rune(passwordUpperCaseLetters)[upperIdx]) +
-		string([]rune(passwordNumbers)[numberIdx]) +
-		string([]rune(passwordLowerCaseLetters)[lowerIdx]) +
-		string([]rune(passwordSpecialChars)[specialIdx])
-
-	for len(password) < minimumLength {
-		i, err := randInt(len(passwordAllChars))
-		if err != nil {
-			return "", err
-		}
-		password = password + string([]rune(passwordAllChars)[i])
-	}
-
-	return password, nil
 }
